moved layers to packages depending on type, reformatted examples to be more compact
