+++ src/main/scala/org/deeplearning4j/scalnet/layers/convolutional/Convolution2D.scala
@@ -18,8 +18,8 @@ package org.deeplearning4j.scalnet.layers.convolutional
 
 import org.deeplearning4j.nn.conf.layers.ConvolutionLayer
 import org.deeplearning4j.nn.weights.WeightInit
-import org.deeplearning4j.scalnet.layers.Layer
-import org.deeplearning4j.scalnet.regularizers.{ NoRegularizer, WeightRegularizer }
+import org.deeplearning4j.scalnet.layers.core.Layer
+import org.deeplearning4j.scalnet.regularizers.{NoRegularizer, WeightRegularizer}
 import org.nd4j.linalg.activations.Activation
 
 /**
