+++ src/main/scala/org/deeplearning4j/scalnet/layers/core/OutputLayer.scala
@@ -14,13 +14,19 @@
  * limitations under the License.
  */
 
-package org.deeplearning4j.scalnet.layers
+package org.deeplearning4j.scalnet.layers.core
 
-import org.nd4j.linalg.lossfunctions.LossFunctions.LossFunction
+import org.deeplearning4j.nn.conf.layers.{OutputLayer => JOutputLayer}
+import org.nd4j.linalg.lossfunctions.LossFunctions
 
 /**
-  * Trait for output layers in DL4J neural networks and computational graphs.
+  * Extension of base layer, used to construct a DL4J OutputLayer after compilation.
+  * OutputLayer has an output object and the ability to return an OutputLayer version
+  * of itself, by providing a loss function.
   *
-  * @author David Kale
+  * @author Max Pumperla
   */
-case class Output(isOutput: Boolean, lossFunction: LossFunction)
+trait OutputLayer extends Layer {
+  def output: Output
+  def toOutputLayer(lossFunction: LossFunctions.LossFunction): OutputLayer
+}
