+++ src/main/scala/org/deeplearning4j/scalnet/layers/pooling/AvgPooling2D.scala
@@ -17,8 +17,8 @@
 package org.deeplearning4j.scalnet.layers.pooling
 
 import org.deeplearning4j.nn.conf.layers.SubsamplingLayer
-import org.deeplearning4j.scalnet.layers.Layer
 import org.deeplearning4j.scalnet.layers.convolutional.Convolution
+import org.deeplearning4j.scalnet.layers.core.Layer
 
 /**
   * 2D average pooling layer in neural net architectures.
