+++ src/main/scala/org/deeplearning4j/scalnet/layers/recurrent/GravesLSTM.scala
@@ -14,11 +14,12 @@
  * limitations under the License.
  */
 
-package org.deeplearning4j.scalnet.layers
+package org.deeplearning4j.scalnet.layers.recurrent
 
 import org.deeplearning4j.nn.conf.layers
 import org.deeplearning4j.nn.weights.WeightInit
-import org.deeplearning4j.scalnet.regularizers.{ NoRegularizer, WeightRegularizer }
+import org.deeplearning4j.scalnet.layers.core.Layer
+import org.deeplearning4j.scalnet.regularizers.{NoRegularizer, WeightRegularizer}
 import org.nd4j.linalg.activations.Activation
 
 class GravesLSTM(nIn: Int,
