+++ src/main/scala/org/deeplearning4j/scalnet/layers/recurrent/RnnOutputLayer.scala
@@ -14,12 +14,13 @@
  * limitations under the License.
  */
 
-package org.deeplearning4j.scalnet.layers
+package org.deeplearning4j.scalnet.layers.recurrent
 
 import org.deeplearning4j.nn.conf.layers
-import org.deeplearning4j.nn.conf.layers.{ OutputLayer => JOutputLayer }
+import org.deeplearning4j.nn.conf.layers.{OutputLayer => JOutputLayer}
 import org.deeplearning4j.nn.weights.WeightInit
-import org.deeplearning4j.scalnet.regularizers.{ NoRegularizer, WeightRegularizer }
+import org.deeplearning4j.scalnet.layers.core.{Layer, Output, OutputLayer}
+import org.deeplearning4j.scalnet.regularizers.{NoRegularizer, WeightRegularizer}
 import org.nd4j.linalg.activations.Activation
 import org.nd4j.linalg.lossfunctions.LossFunctions
 import org.nd4j.linalg.lossfunctions.LossFunctions.LossFunction
