+++ src/main/scala/org/deeplearning4j/scalnet/models/Sequential.scala
@@ -17,10 +17,11 @@
 package org.deeplearning4j.scalnet.models
 
 import org.deeplearning4j.nn.api.OptimizationAlgorithm
-import org.deeplearning4j.nn.conf.{ NeuralNetConfiguration, Updater }
+import org.deeplearning4j.nn.conf.{NeuralNetConfiguration, Updater}
 import org.deeplearning4j.nn.multilayer.MultiLayerNetwork
 import org.deeplearning4j.optimize.api.IterationListener
 import org.deeplearning4j.scalnet.layers._
+import org.deeplearning4j.scalnet.layers.core.{Layer, Node, Preprocessor}
 import org.deeplearning4j.scalnet.logging.Logging
 import org.nd4j.linalg.dataset.api.DataSet
 import org.nd4j.linalg.dataset.api.iterator.DataSetIterator
