+++ src/test/scala/org/deeplearning4j/scalnet/examples/keras/convolution/LeNetMnistExample.scala
@@ -17,12 +17,11 @@
 package org.deeplearning4j.scalnet.examples.keras.convolution
 
 import org.deeplearning4j.datasets.iterator.impl.MnistDataSetIterator
-import org.deeplearning4j.nn.weights.WeightInit
 import org.deeplearning4j.optimize.listeners.ScoreIterationListener
-import org.deeplearning4j.scalnet.layers.Dense
 import org.deeplearning4j.scalnet.layers.convolutional.Convolution2D
+import org.deeplearning4j.scalnet.layers.core.Dense
 import org.deeplearning4j.scalnet.layers.pooling.MaxPooling2D
-import org.deeplearning4j.scalnet.layers.reshaping.{ Flatten3D, Unflatten3D }
+import org.deeplearning4j.scalnet.layers.reshaping.{Flatten3D, Unflatten3D}
 import org.deeplearning4j.scalnet.logging.Logging
 import org.deeplearning4j.scalnet.models.Sequential
 import org.deeplearning4j.scalnet.regularizers.L2
@@ -42,6 +41,7 @@ object LeNetMnistExample extends App with Logging {
   val width: Int = 28
   val channels: Int = 1
   val nClasses: Int = 10
+
   val batchSize: Int = 64
   val epochs: Int = 10
   val weightDecay: Double = 0.0005
@@ -53,32 +53,18 @@ object LeNetMnistExample extends App with Logging {
 
   logger.info("Build model...")
   val model: Sequential = Sequential(rngSeed = seed)
-  model.add(Unflatten3D(List(height, width, channels), nIn = height * width))
-  model.add(
-    Convolution2D(nFilter = 20,
-                  kernelSize = List(5, 5),
-                  stride = List(1, 1),
-                  weightInit = WeightInit.XAVIER,
-                  regularizer = L2(weightDecay),
-                  activation = Activation.RELU)
-  )
-  model.add(MaxPooling2D(kernelSize = List(2, 2), stride = List(2, 2)))
-  model.add(
-    Convolution2D(nFilter = 50,
-                  kernelSize = List(5, 5),
-                  stride = List(1, 1),
-                  weightInit = WeightInit.XAVIER,
-                  regularizer = L2(weightDecay),
-                  activation = Activation.RELU)
-  )
-  model.add(MaxPooling2D(kernelSize = List(2, 2), stride = List(2, 2)))
-  model.add(Flatten3D())
-  model.add(
-    Dense(nOut = 512, weightInit = WeightInit.XAVIER, activation = Activation.RELU, regularizer = L2(weightDecay))
-  )
-  model.add(Dense(nClasses, weightInit = WeightInit.XAVIER, activation = Activation.SOFTMAX))
 
-  model.compile(lossFunction = LossFunction.NEGATIVELOGLIKELIHOOD)
+  model.add(Unflatten3D(List(height, width, channels), nIn = height * width))
+  model.add(Convolution2D(20, List(5, 5), channels, regularizer = L2(weightDecay), activation = Activation.RELU))
+  model.add(MaxPooling2D(List(2, 2), List(2, 2)))
+
+  model.add(Convolution2D(50, List(5, 5), regularizer = L2(weightDecay), activation = Activation.RELU))
+  model.add(MaxPooling2D(List(2, 2),  List(2, 2)))
+  model.add(Flatten3D())
+
+  model.add(Dense(512, regularizer = L2(weightDecay), activation = Activation.RELU))
+  model.add(Dense(nClasses, activation = Activation.SOFTMAX))
+  model.compile(LossFunction.NEGATIVELOGLIKELIHOOD)
 
   logger.info("Train model...")
   model.fit(mnistTrain, epochs, List(new ScoreIterationListener(scoreFrequency)))
