+++ src/test/scala/org/deeplearning4j/scalnet/layers/embeddings/EmbeddingLayerTest.scala
@@ -14,30 +14,30 @@
  * limitations under the License.
  */
 
-package org.deeplearning4j.scalnet.layers.LSTM
+package org.deeplearning4j.scalnet.layers.embeddings
 
-import org.deeplearning4j.scalnet.layers.GravesLSTM
-import org.scalatest.{ Matchers, WordSpec }
+import org.scalatest.{Matchers, WordSpec}
 
-class GravesLSTMTest extends WordSpec with Matchers {
+class EmbeddingLayerTest extends WordSpec with Matchers {
 
-  "A Graves LSTM layer" should {
+  "An embedding layer" should {
 
     "have an input layer of shape (10, 100)" in {
-      val gravesLSTMLayer = GravesLSTM(10, 100)
-      gravesLSTMLayer.inputShape shouldBe List(10, 100)
+      val embeddingLayer = EmbeddingLayer(10, 100)
+      embeddingLayer.inputShape shouldBe List(10, 100)
     }
 
     "have an ouput layer of shape (10, 100)" in {
-      val gravesLSTMLayer = GravesLSTM(10, 100)
-      gravesLSTMLayer.outputShape shouldBe List(100, 10)
+      val embeddingLayer = EmbeddingLayer(10, 100)
+      embeddingLayer.outputShape shouldBe List(100, 10)
     }
 
-    "compile to a DL4J GravesLSTM" in {
-      val gravesLSTMLayer = GravesLSTM(10, 100)
-      val compiledLayer = gravesLSTMLayer.compile
-      compiledLayer.isInstanceOf[org.deeplearning4j.nn.conf.layers.GravesLSTM] shouldBe true
+    "compile to a DL4J EmbeddingLayer" in {
+      val embeddingLayer = EmbeddingLayer(10, 100)
+      val compiledLayer = embeddingLayer.compile
+      compiledLayer.isInstanceOf[org.deeplearning4j.nn.conf.layers.EmbeddingLayer] shouldBe true
     }
 
   }
+
 }
