+++ src/test/scala/org/deeplearning4j/scalnet/layers/recurrent/GravesLSTMTest.scala
@@ -14,30 +14,29 @@
  * limitations under the License.
  */
 
-package org.deeplearning4j.scalnet.layers
+package org.deeplearning4j.scalnet.layers.recurrent
 
-import org.scalatest.{ Matchers, WordSpec }
+import org.scalatest.{Matchers, WordSpec}
 
-class EmbeddingLayerTest extends WordSpec with Matchers {
+class GravesLSTMTest extends WordSpec with Matchers {
 
-  "An embedding layer" should {
+  "A Graves LSTM layer" should {
 
     "have an input layer of shape (10, 100)" in {
-      val embeddingLayer = EmbeddingLayer(10, 100)
-      embeddingLayer.inputShape shouldBe List(10, 100)
+      val gravesLSTMLayer = GravesLSTM(10, 100)
+      gravesLSTMLayer.inputShape shouldBe List(10, 100)
     }
 
     "have an ouput layer of shape (10, 100)" in {
-      val embeddingLayer = EmbeddingLayer(10, 100)
-      embeddingLayer.outputShape shouldBe List(100, 10)
+      val gravesLSTMLayer = GravesLSTM(10, 100)
+      gravesLSTMLayer.outputShape shouldBe List(100, 10)
     }
 
-    "compile to a DL4J EmbeddingLayer" in {
-      val embeddingLayer = EmbeddingLayer(10, 100)
-      val compiledLayer = embeddingLayer.compile
-      compiledLayer.isInstanceOf[org.deeplearning4j.nn.conf.layers.EmbeddingLayer] shouldBe true
+    "compile to a DL4J GravesLSTM" in {
+      val gravesLSTMLayer = GravesLSTM(10, 100)
+      val compiledLayer = gravesLSTMLayer.compile
+      compiledLayer.isInstanceOf[org.deeplearning4j.nn.conf.layers.GravesLSTM] shouldBe true
     }
 
   }
-
 }
