+++ src/test/scala/org/deeplearning4j/scalnet/models/NeuralNetTest.scala
@@ -16,9 +16,9 @@
 
 package org.deeplearning4j.scalnet.models
 
-import org.deeplearning4j.scalnet.layers.{ Dense, OutputLayer }
+import org.deeplearning4j.scalnet.layers.core.{Dense, OutputLayer}
 import org.nd4j.linalg.lossfunctions.LossFunctions.LossFunction
-import org.scalatest.{ BeforeAndAfter, FunSpec }
+import org.scalatest.{BeforeAndAfter, FunSpec}
 
 /**
   * Created by maxpumperla on 19/07/17.
